Started implementing rolling ball operation

+++ src/Kernel/RgbPixel.cs
@@ -1,9 +1,9 @@
 ﻿namespace Kernel {
 
     public class RgbPixel {
-        public byte Red;
-        public byte Green;
-        public byte Blue;
+        public short Red;
+        public short Green;
+        public short Blue;
     }
 
 }
